- WorkingDayMaster SF 마이그레이션 프레임워크 보강 (common.kts/verify/README) + owner/audit FK
- 앱 패키지 다운로드 URL 복사 시 만료 시각 함께 복사

+++ backend/src/main/kotlin/com/otoki/powersales/sales/repository/WorkingDayMasterRepository.kt
@@ -32,6 +32,6 @@ interface WorkingDayMasterRepository : JpaRepository<WorkingDayMaster, Long> {
     fun countWorkingDays(
         @Param("start") start: LocalDate,
         @Param("end") end: LocalDate,
-        @Param("check") check: Int,
+        @Param("check") check: Double,
     ): Long
 }
